+++ dtaianomaly/__init__.py
@@ -5,4 +5,3 @@ from . import pipeline
 from . import thresholding
 from . import workflow
 from . import anomaly_detection
-