+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Added option ``novelty`` to ``MatrixProfileDetector``, which will compute the
   matrix profile in regard to the train data, if ``novelty=True``. By default, 
   the matrix profile is computed based on a self-join of the test data. 
+- Implemented ``MedianMethod`` anomaly detector. 
 
 ### Changed